requirements.txt
Django
python-dotenv
pillow
langchain
langchain-community
langchain-core
google-generativeai 
langchain-google-genai
temp